+++ packages/plugin-hardhat/test/uups-happy-path.js
@@ -13,7 +13,7 @@ test('happy path', async t => {
 
   const greeter = await upgrades.deployProxy(Greeter, ['Hello, Hardhat!'], { kind: 'uups' });
 
-  const greeter2 = await upgrades.upgradeProxy(greeter.address, GreeterV2);
+  const greeter2 = await upgrades.upgradeProxy(greeter, GreeterV2);
   await greeter2.resetGreeting();
 
   const greeter3ImplAddr = await upgrades.prepareUpgrade(greeter.address, GreeterV3);