+++ packages/plugin-truffle/src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './options';
 export * from './truffle';
 export * from './validations';
 export * from './wrap-provider';
+export * from './contract-address';